clean code work

+++ src/app/join/join-header/join-header.component.ts
@@ -17,17 +17,22 @@ export class JoinHeaderComponent {
     private as: AuthService,
     private router: Router,
     public menuService: MenuService
-  ) { }
+  ) {
+    this.menuService.headerSubmenuView = false;
+   }
 
 
+  /**
+   * Initiates the logout process by navigating to the login page, logging out the user, and clearing local storage upon successful logout.
+   * 
+   * @returns {Promise<void>} A Promise that resolves when the logout process is completed.
+   */
   async logout() {
     try {
       this.router.navigateByUrl('/login');
       let response = await this.as.logout();
       console.log(response.message);
-      if (response.message == 'logout successfully') {
-        localStorage.clear();
-      }
+      if (response.message == 'logout successfully') localStorage.clear();
     } catch (err) {
       alert('logout error');
     }
